==> 3. Data Merger.ipynb <==
import pandas as pd
import numpy as np
import glob
import os

joined_files = os.path.join("Data *.csv")
joined_list = glob.glob(joined_files)
train = pd.concat(map(pd.read_csv, joined_list), ignore_index=True)
print(train.columns.tolist())
#train.drop(['Unnamed: 0', 'Unnamed: 0', '0_FFT mean coefficient_0', '0_FFT mean coefficient_1', '0_FFT mean coefficient_2', '0_FFT mean coefficient_3', '0_FFT mean coefficient_4', '0_FFT mean coefficient_5', '0_FFT mean coefficient_6', '0_FFT mean coefficient_7', '0_FFT mean coefficient_8', '0_FFT mean coefficient_9', '0_FFT mean coefficient_10', '0_FFT mean coefficient_11', '0_FFT mean coefficient_12', '0_FFT mean coefficient_13', '0_FFT mean coefficient_14', '0_FFT mean coefficient_15', '0_FFT mean coefficient_16', '0_FFT mean coefficient_17', '0_FFT mean coefficient_18', '0_FFT mean coefficient_19', '0_FFT mean coefficient_20', '0_FFT mean coefficient_21', '0_FFT mean coefficient_22', '0_FFT mean coefficient_23', '0_FFT mean coefficient_24', '0_FFT mean coefficient_25', '0_FFT mean coefficient_26', '0_FFT mean coefficient_27', '0_FFT mean coefficient_28', '0_FFT mean coefficient_29', '0_FFT mean coefficient_30', '0_FFT mean coefficient_31', '0_FFT mean coefficient_32', '0_FFT mean coefficient_33', '0_FFT mean coefficient_34', '0_FFT mean coefficient_35', '0_FFT mean coefficient_36', '0_FFT mean coefficient_37', '0_FFT mean coefficient_38', '0_FFT mean coefficient_39', '0_FFT mean coefficient_40', '0_FFT mean coefficient_41', '0_FFT mean coefficient_42', '0_FFT mean coefficient_43', '0_FFT mean coefficient_44', '0_FFT mean coefficient_45', '0_FFT mean coefficient_46', '0_FFT mean coefficient_47', '0_FFT mean coefficient_48', '0_FFT mean coefficient_49', '0_FFT mean coefficient_50', '0_Fundamental frequency', '0_Human range energy', '0_LPCC_0', '0_LPCC_1', '0_LPCC_2', '0_LPCC_3', '0_LPCC_4', '0_LPCC_5', '0_LPCC_6', '0_LPCC_7', '0_LPCC_8', '0_LPCC_9', '0_LPCC_10', '0_LPCC_11', '0_LPCC_12', '0_MFCC_0', '0_MFCC_1', '0_MFCC_2', '0_MFCC_3', '0_MFCC_4', '0_MFCC_5', '0_MFCC_6', '0_MFCC_7', '0_MFCC_8', '0_MFCC_9', '0_MFCC_10','0_MFCC_11','0_FFT mean coefficient_51', '0_FFT mean coefficient_52', '0_FFT mean coefficient_53', '0_FFT mean coefficient_54', '0_FFT mean coefficient_55', '0_FFT mean coefficient_56', '0_FFT mean coefficient_57', '0_FFT mean coefficient_58', '0_FFT mean coefficient_59', '0_FFT mean coefficient_60', '0_FFT mean coefficient_61', '0_FFT mean coefficient_62', '0_FFT mean coefficient_63', '0_FFT mean coefficient_64', '0_FFT mean coefficient_65', '0_FFT mean coefficient_66', '0_FFT mean coefficient_67', '0_FFT mean coefficient_68', '0_FFT mean coefficient_69', '0_FFT mean coefficient_70', '0_FFT mean coefficient_71', '0_FFT mean coefficient_72', '0_FFT mean coefficient_73', '0_FFT mean coefficient_74', '0_FFT mean coefficient_75', '0_FFT mean coefficient_76', '0_FFT mean coefficient_77', '0_FFT mean coefficient_78', '0_FFT mean coefficient_79', '0_FFT mean coefficient_80', '0_FFT mean coefficient_81', '0_FFT mean coefficient_82', '0_FFT mean coefficient_83', '0_FFT mean coefficient_84', '0_FFT mean coefficient_85', '0_FFT mean coefficient_86', '0_FFT mean coefficient_87', '0_FFT mean coefficient_88', '0_FFT mean coefficient_89', '0_FFT mean coefficient_90', '0_FFT mean coefficient_91', '0_FFT mean coefficient_92', '0_FFT mean coefficient_93', '0_FFT mean coefficient_94', '0_FFT mean coefficient_95', '0_FFT mean coefficient_96', '0_FFT mean coefficient_97', '0_FFT mean coefficient_98', '0_FFT mean coefficient_99', '0_FFT mean coefficient_100','0_FFT mean coefficient_101', '0_FFT mean coefficient_102', '0_FFT mean coefficient_103', '0_FFT mean coefficient_104', '0_FFT mean coefficient_105', '0_FFT mean coefficient_106', '0_FFT mean coefficient_107', '0_FFT mean coefficient_108', '0_FFT mean coefficient_109', '0_FFT mean coefficient_110', '0_FFT mean coefficient_111', '0_FFT mean coefficient_112', '0_FFT mean coefficient_113', '0_FFT mean coefficient_114', '0_FFT mean coefficient_115', '0_FFT mean coefficient_116', '0_FFT mean coefficient_117', '0_FFT mean coefficient_118', '0_FFT mean coefficient_119', '0_FFT mean coefficient_120', '0_FFT mean coefficient_121', '0_FFT mean coefficient_122', '0_FFT mean coefficient_123', '0_FFT mean coefficient_124', '0_FFT mean coefficient_125', '0_FFT mean coefficient_126', '0_FFT mean coefficient_127', '0_FFT mean coefficient_128', '0_FFT mean coefficient_129', '0_FFT mean coefficient_130', '0_FFT mean coefficient_131', '0_FFT mean coefficient_132', '0_FFT mean coefficient_133', '0_FFT mean coefficient_134', '0_FFT mean coefficient_135', '0_FFT mean coefficient_136', '0_FFT mean coefficient_137', '0_FFT mean coefficient_138', '0_FFT mean coefficient_139', '0_FFT mean coefficient_140', '0_FFT mean coefficient_141', '0_FFT mean coefficient_142', '0_FFT mean coefficient_143', '0_FFT mean coefficient_144', '0_FFT mean coefficient_145', '0_FFT mean coefficient_146', '0_FFT mean coefficient_147', '0_FFT mean coefficient_148', '0_FFT mean coefficient_149', '0_FFT mean coefficient_150','0_FFT mean coefficient_151', '0_FFT mean coefficient_152', '0_FFT mean coefficient_153', '0_FFT mean coefficient_154', '0_FFT mean coefficient_155', '0_FFT mean coefficient_156', '0_FFT mean coefficient_157', '0_FFT mean coefficient_158', '0_FFT mean coefficient_159', '0_FFT mean coefficient_160', '0_FFT mean coefficient_161', '0_FFT mean coefficient_162', '0_FFT mean coefficient_163', '0_FFT mean coefficient_164', '0_FFT mean coefficient_165', '0_FFT mean coefficient_166', '0_FFT mean coefficient_167', '0_FFT mean coefficient_168', '0_FFT mean coefficient_169', '0_FFT mean coefficient_170', '0_FFT mean coefficient_171', '0_FFT mean coefficient_172', '0_FFT mean coefficient_173', '0_FFT mean coefficient_174', '0_FFT mean coefficient_175', '0_FFT mean coefficient_176', '0_FFT mean coefficient_177', '0_FFT mean coefficient_178', '0_FFT mean coefficient_179', '0_FFT mean coefficient_180', '0_FFT mean coefficient_181', '0_FFT mean coefficient_182', '0_FFT mean coefficient_183', '0_FFT mean coefficient_184', '0_FFT mean coefficient_185', '0_FFT mean coefficient_186', '0_FFT mean coefficient_187', '0_FFT mean coefficient_188', '0_FFT mean coefficient_189', '0_FFT mean coefficient_190', '0_FFT mean coefficient_191', '0_FFT mean coefficient_192', '0_FFT mean coefficient_193', '0_FFT mean coefficient_194', '0_FFT mean coefficient_195', '0_FFT mean coefficient_196', '0_FFT mean coefficient_197', '0_FFT mean coefficient_198', '0_FFT mean coefficient_199', '0_FFT mean coefficient_200'], axis=1, inplace=True)
train.to_csv(r'Train 400.csv')
train
